requirements.txt
langchain
langchain-community
faiss-cpu
openai
tiktoken
InstructorEmbedding
sentence-transformers
streamlit
python-dotenv
pandas
trafilatura
